% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgcvTrans.R
\docType{package}
\name{mgcvTrans}
\alias{mgcvTrans}
\alias{mgcvTrans-package}
\title{mgcvTrans: extra basis definitions for the \code{mgcv} package}
\description{
This package defines a handful of extra bases for the \code{mgcv} package
involving transformation of variables.
}

