% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_basis.R
\name{Predict.matrix.ar.smooth}
\alias{Predict.matrix.ar.smooth}
\title{Predict.matrix method for ar basis}
\usage{
\method{Predict.matrix}{ar.smooth}(object, data)
}
\arguments{
\item{object}{a \code{ar.smooth} object created by
\code{\link{smooth.construct.ar.smooth.spec}}, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for \code{dt} terms
}
\description{
Predict.matrix method for ar basis
}
\author{
Jonathan Gellar
}
