% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dt_basis.R
\name{Predict.matrix.dt.smooth}
\alias{Predict.matrix.dt.smooth}
\title{Predict.matrix method for dt basis}
\usage{
Predict.matrix.dt.smooth(object, data)
}
\description{
Predict.matrix method for dt basis
}

