% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nb_basis.R
\name{Predict.matrix.nb.smooth}
\alias{Predict.matrix.nb.smooth}
\title{Predict.matrix method for nb basis}
\usage{
Predict.matrix.nb.smooth(object, data)
}
\description{
Predict.matrix method for nb basis
}

