% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_basis.R
\name{smooth.construct.ar.smooth.spec}
\alias{smooth.construct.ar.smooth.spec}
\title{Autoregressive Penalty}
\usage{
smooth.construct.ar.smooth.spec(object, data, knots)
}
\description{
Autoregressive Penalty
}
