% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smooth.construct.dt.smooth.spec.R
\name{Predict.matrix.dt.smooth}
\alias{Predict.matrix.dt.smooth}
\title{Predict.matrix method for dt basis}
\usage{
\method{Predict.matrix}{dt.smooth}(object, data)
}
\description{
Predict.matrix method for dt basis
}

