% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nb_basis.R
\name{Predict.matrix.nb.smooth}
\alias{Predict.matrix.nb.smooth}
\title{Predict.matrix method for nb basis}
\usage{
\method{Predict.matrix}{nb.smooth}(object, data)
}
\arguments{
\item{object}{a \code{nb.smooth} object created by
\code{\link{smooth.construct.nb.smooth.spec}}, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for \code{dt} terms
}
\description{
Predict.matrix method for nb basis
}
\author{
Jonathan Gellar
}

