% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_basis.R
\name{smooth.construct.ar.smooth.spec}
\alias{smooth.construct.ar.smooth.spec}
\title{Autoregressive Penalty}
\usage{
\method{smooth.construct}{ar.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, generated by \code{s()},
\code{te()}, \code{ti()}, or \code{t2()}, with \code{bs="ar"}. The order of
the autoregressive penalty is specified using the \code{m} argument, with
the default being AR(2).}

\item{data}{a list containing just the data (including any by variable)
required by this term, with names corresponding to \code{object$term}
(and \code{object$by}). The \code{by} variable is the last element.}

\item{knots}{a list containing any knots supplied for basis setup - in same
order and with same names as \code{data}. If supplied, coefficients
will be used only at the indicated knot values. If \code{NULL},
all observed data values will be used.}
}
\description{
The \code{ar} "basis" uses a separate coefficient for each value of
the variables this is a smooth function of. The coefficients are penalized
with an autoregressive penalty. This basis is useful when there are not very
many unique values of the variable, making dimension reduction unnecessary.
}
\details{
Note that the predict method needs to have the ability to evaluate the
basis at any point in the range of the data; this requires interpolation
for values between the observed data. The default is to use linear
interpolation to evaluate the basis at these points. Alternative
interpolation methods may be specified using the \code{xt} argument of the
function used to create \code{object} (usually \code{s()}). \code{xt="lowess"}
indicates lowess smoothing, and \code{xt="bspline"} indicates (unpenalized)
b-spline smoothing. This method will also be used in the construction of the
basis if the user specifies knots that are not equal to the unique data values.
}
\seealso{
\code{\link[mgcv]{smooth.construct}}
}
\author{
Jonathan Gellar \email{JGellar@mathematica-mpr.com}
}
