% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nb_basis.R
\name{smooth.construct.nb.smooth.spec}
\alias{smooth.construct.nb.smooth.spec}
\title{Smooths with No Basis}
\usage{
smooth.construct.nb.smooth.spec(object, data, knots)
}
\description{
The \code{nb} "basis" uses a separate coefficient for each value of
the variables this is a smooth function of. The coefficients are penalized
with a difference penalty. This basis is most useful when there are not very
many unique values of the variable, making dimension reduction unnecessary.
}
\examples{
N <- 100
x <- sample(-3:3, N, replace=T)
y <- sin(2*pi*x/6) + rnorm(N, sd=.5)
mod <- gam(y ~ s(x, bs="nb"))
pre <- predict(mod, newdata=data.frame(x=c(-3:3)))
pre <- predict(mod, newdata=data.frame(x=seq(-3,3,length=100)))
mod <- gam(y ~ s(x, bs="nb", xt="lowess"))
pre <- predict(mod, newdata=data.frame(x=seq(-3,3,length=100)))
}

